materialize-bigquery: batch table creation and run batches concurrently
Creating hundreds of tables in a single apply operation takes quite a while when run using a single
statement. This change creates batches of table creation statement and runs them with concurrent
workers.

BigQuery does not support DDL statements in transactions, so there's no reason to run them as part
of a transaction. We just need to make sure the materialization spec update happens only after all
of the table alterations are complete.

With this strategy, we can create several hundred tables within the 5 minute timeout period allowed
for an Apply RPC.